updated the updater

+++ cogs/update.py
@@ -13,7 +13,13 @@ import config
 import globalconfig
 import importlib
 #import socket
-import wget 
+import wget
+import time
+import re
+from zipfile import ZipFile
+import versionfile
+import requests
+from datetime import datetime
 
 '''def close_port():
         message = 'disconnect'
@@ -32,66 +38,26 @@ class Update(commands.Cog):
     @commands.command()
     async def updatecheck(self, ctx):
         if str(ctx.message.author.id) == config.ownerID:
-            em = discord.Embed(title = "This command is in development", color = discord.Color.red())
-            await ctx.send(embed = em)
-        #     # username = os.getlogin()
-        #     if sys.platform == "linux" or sys.platform == "linux2" or sys.platform == "darwin":
-        #         tmpdir = "/tmp"
-        #     elif sys.platform == "win32":
-        #         tmpdir = "/Temp"
-        #     with open('config.py') as f:
-        #         if not 'latest_version' in f.read():
-        #             with open('config.py', 'a') as writeFile :
-        #                 writeFile.write("latest_version = 'unknown'")
-        #                 writeFile.close()
-        #                 importlib.reload(config)
-        #     if not os.path.exists(tmpdir + '/updatecheck'):
-        #         os.makedirs(tmpdir + '/updatecheck')
-        #     elif os.path.exists(tmpdir + '/updatecheck'):
-        #         if os.path.exists(tmpdir + '/updatecheck/.git/objects/pack'):
-        #             new_name = str("unlock")
-        #             os.rename(tmpdir + '/updatecheck/.git/objects/pack', new_name)
-        #             shutil.rmtree('unlock')
-        #         shutil.rmtree(tmpdir + '/updatecheck')
-        #     #os.mkdir('/tmp/freeupdate')
-        #     HTTPS_REMOTE_URL = globalconfig.github_login_url
-        #     first_embed = discord.Embed(title = "Checking for updates...", description = "FreeDiscord is now checking for updates. Please be patient.", color = discord.Color.orange())
-        #     # send a first message with an embed
-        #     msg = await ctx.send(embed=first_embed)
-        #     DEST_NAME = tmpdir + '/updatecheck'
-        #     cloned_repo = Repo.clone_from(HTTPS_REMOTE_URL, DEST_NAME)
-        #     dir_path = os.getcwd()
-        #     copyfile(tmpdir + '/updatecheck/globalconfig.py', dir_path + '/updateconfig.py')
-        #     try:
-        #         shutil.rmtree(tmpdir + '/updatecheck')
-        #     except os.error:
-        #         embed = discord.Embed(title = "Error in removing `" + tmpdir + "/updatecheck` folder", description = 'The `' + tmpdir + '/updatecheck` folder was not able to be removed, probably due to a permissions issue.', color = discord.Color.red())
-        #         await ctx.send(embed=embed) 
-        #     import updateconfig
-        #     if updateconfig.version > globalconfig.version:
-        #         new_embed = discord.Embed(title = "Checking for updates...", description = "Checking for updates succeeded!", color = discord.Color.green())
-        #         new_embed.add_field(name = "Upgrade found!", value = "It is recommended to update to version " + updateconfig.version + " from version " + globalconfig.version + " for the latest bug fixes and feature improvements.")
-        #         new_embed.add_field(name = "How do I upgrade?", value = "Use `" + config.prefix + "help updatebot` for more details.")
-        #         await msg.edit(embed=new_embed)
-        #     if updateconfig.version < globalconfig.version:
-        #         new_embed = discord.Embed(title = "Checking for updates...", description = "Checking for updates succeeded!", color = discord.Color.green())
-        #         new_embed.add_field(name = "Downgrade found!", value = "It is recommended to downgrade to version " + updateconfig.version + " from version " + globalconfig.version + " because something most likely broke in the latest release.")
-        #         new_embed.add_field(name = "How do I downgrade?", value = "Use `" + config.prefix + "help updatebot` for more details. (The update command also downgrades the bot.)")
-        #         await msg.edit(embed=new_embed)
-        #     if updateconfig.version == globalconfig.version:
-        #         new_embed = discord.Embed(title = "Checking for updates...", description = "Checking for updates succeeded!", color = discord.Color.green())
-        #         new_embed.add_field(name = "No updates found!", value = "You are up to date! This bot is at version `" + globalconfig.version + "` and the latest bot files available are at version `" + updateconfig.version + "`.")
-        #         new_embed.add_field(name = "How do I upgrade?", value = "You don't need to take any action, as you are up to date already. However, you can use `" + config.prefix + "help updatebot` for more details about the upgrade/downgrade process.")
-        #         await msg.edit(embed=new_embed)
-        #     with open('config.py', 'r') as file :
-        #         filedata = file.read()
-        #     newdata = filedata.replace(config.latest_version, updateconfig.version)
-        #     with open('config.py', 'w') as file:
-        #         file.write(newdata)
-        #     file.close()
-        #     importlib.reload(config)
-        #     os.remove(dir_path + "/updateconfig.py")
-
+            firstem = discord.Embed(title = "FOSSDiscord is checking for updates...", color = discord.Color.orange())
+            embedmsg = await ctx.send(embed=firstem)
+            latestversionresponse = requests.get("https://api.github.com/repos/FOSS-Devs/fossdiscord/releases/latest")
+            latestversionget = latestversionresponse.json()["name"]
+            latestversion = latestversionget.split(' ', 1)[1]
+            if versionfile.currentversion == latestversion:
+                secondem = discord.Embed(title = "`Updatecheck`", color = discord.Color.green())
+                secondem.add_field(name = "Checking for updates succeeded!", value = "There are no updates available.")
+                await embedmsg.edit(embed=secondem)
+            elif versionfile.currentversion > latestversion:
+                secondem = discord.Embed(title = "`Updatecheck`", color = discord.Color.green())
+                secondem.add_field(name = "Invalid version in the versionfile.")
+                secondem.add_field(name = "There is an invalid version in the versionfile, try downloading a fresh copy of FOSSDiscord.", color = discord.Color.green())
+                await embedmsg.edit(embed=secondem)
+            else:
+                secondem = discord.Embed(title = f"\n \n", color = discord.Color.green())
+                secondem.add_field(name = "Checking for updates succeeded!")
+                secondem.add_field(name = f"You can update the bot from {versionfile.currentversion} to {latestversion}.")
+                secondem.add_field(name = f"You can update the bot with {config.prefix}updatebot.")
+                await embedmsg.edit(embed=secondem)
         else:
             em = discord.Embed(title = "This command is for the bot owner only.", color = discord.Color.red())
             await ctx.send(embed = em)
@@ -100,43 +66,51 @@ class Update(commands.Cog):
     async def updatebot(self, ctx):
         if str(ctx.message.author.id) == config.ownerID:
            if sys.platform == "linux" or sys.platform == "linux2" or sys.platform == "win32":
-                datetimenow = datetime.now()
-                currentdate = datetime.strftime(datetimenow, '%a %b %y %I%M%p')
+                if versionfile.currentversion == latestversion:
+                    em = discord.Embed(title = "Already the latest version", color = discord.Color.green())
+                    em.add_field(name = '', value = "There are no updates available.")
+                    await ctx.send(embed=em)
+                elif versionfile.currentversion > latestversion:
+                    em = discord.Embed(title = "Invalid version in the versionfile.", color = discord.Color.green())
+                    em.add_field(name = '', value = "There is an invalid version in the versionfile, try downloading a fresh copy of FOSSDiscord.")
+                    await ctx.send(embed=em)
+                else:
+                    datetimenow = datetime.now()
+                    currentdate = datetime.strftime(datetimenow, '%a %b %y %I%M%p')
 
-                latestversionresponse = requests.get("https://api.github.com/repos/FOSS-Devs/fossdiscord/releases/latest")
-                latestversionget = latestversionresponse.json()["name"]
-                latestversion = latestversionget.split(' ', 1)[1]
+                    latestversionresponse = requests.get("https://api.github.com/repos/FOSS-Devs/fossdiscord/releases/latest")
+                    latestversionget = latestversionresponse.json()["name"]
+                    latestversion = latestversionget.split(' ', 1)[1]
 
-                downloadurl = f"https://codeload.github.com/FOSS-Devs/fossdiscord/zip/v{latestversion}"
-                currentdir = os.getcwd()
+                    downloadurl = f"https://codeload.github.com/FOSS-Devs/fossdiscord/zip/v{latestversion}"
+                    currentdir = os.getcwd()
 
-                os.mkdir(f'{currentdir}/updatertemp')
-                os.mkdir(f'{currentdir}/updatertemp/unzipped')
-                wget.download(downloadurl, f'{currentdir}/updatertemp/download.zip')
+                    os.mkdir(f'{currentdir}/updatertemp')
+                    os.mkdir(f'{currentdir}/updatertemp/unzipped')
+                    wget.download(downloadurl, f'{currentdir}/updatertemp/download.zip')
 
-                with ZipFile(f'{currentdir}/updatertemp/download.zip', 'r') as zipObj:
-                    zipObj.extractall(f'{currentdir}/updatertemp/unzipped')
+                    with ZipFile(f'{currentdir}/updatertemp/download.zip', 'r') as zipObj:
+                        zipObj.extractall(f'{currentdir}/updatertemp/unzipped')
 
-                os.mkdir(f'{currentdir}/backup-{currentdate}')
+                    os.mkdir(f'{currentdir}/backup-{currentdate}')
 
-                allfiles = ["bot.py", "setup.py", "cogs", "docs", "src", "README.md", ".gitignore", "LICENSE", "globalconfig.py"]
+                    allfiles = ["bot.py", "setup.py", "cogs", "docs", "src", "README.md", ".gitignore", "LICENSE", "globalconfig.py"]
 
-                for f in allfiles:
-                    shutil.move(f'{currentdir}/' + f, f'{currentdir}/backup-{currentdate}')
+                    for f in allfiles:
+                        shutil.move(f'{currentdir}/' + f, f'{currentdir}/backup-{currentdate}')
 
-                allfilesunzipped = os.listdir(f'{currentdir}/updatertemp/unzipped/fossdiscord-{latestversion}')
+                    allfilesunzipped = os.listdir(f'{currentdir}/updatertemp/unzipped/fossdiscord-{latestversion}')
 
-                for f in allfilesunzipped:
-                    shutil.move(f'{currentdir}/updatertemp/unzipped/fossdiscord-{latestversion}/' + f, currentdir)
+                    for f in allfilesunzipped:
+                        shutil.move(f'{currentdir}/updatertemp/unzipped/fossdiscord-{latestversion}/' + f, currentdir)
 
-                shutil.rmtree("updatertemp")
-                print("Done! Restart the bot to apply the changes!")
-                em = discord.Embed(title = "Updated!", description = "FOSSDiscord updated! No error reported. Check your console to confirm this.", color = discord.Color.green())
-                em.add_field(name = "Note", value = "You have to start the bot again manually. If it won't start, open an issue in FOSSDiscord's GitHub repository.")
-                await ctx.send(embed = em)
-                #close_port()
-                await ctx.bot.close()
-                subprocess.Popen([f'python3 {currentdir}/bot.py'])
+                    shutil.rmtree("updatertemp")
+                    print("Done! Restart the bot to apply the changes!")
+                    em = discord.Embed(title = "Updated!", description = "FOSSDiscord updated! No error reported. Check your console to confirm this.", color = discord.Color.green())
+                    em.add_field(name = "Note", value = "You have to start the bot again manually. If it won't start, open an issue in FOSSDiscord's GitHub repository.")
+                    await ctx.send(embed = em)
+                    #close_port()
+                    await ctx.bot.close()
            elif sys.platform == "darwin":
                 em = discord.Embed(title = "We are still testing `updatebot` for macOS.", color = discord.Color.red())
                 await ctx.send(embed = em)
